feat(list): Phase 4 — on-demand virtualization via Virtualizer<T,K>
Replace the eager-materialize __h() with a pull-based core:

- add whisker-runtime::view::virtualize<T,K> — the container-agnostic
  Virtualizer (whisker#83: ListMount = Virtualizer + <list>). Builds
  items lazily in componentAtIndex under a per-slot owner, stamps a
  STABLE item-key (key->id map, reorder-safe), recycles on
  enqueueComponent, bumps layout-id per data version.
- __h() now just calls virtualize(); only visible slots hold elements
  (was: every item materialized up front).
- list finaliser gains a `T: Clone` bound (items cloned per slot).
- remove the unused list_mount v1; virtualizer subsumes it (+ tests for
  stable-item-key-across-reorder, enqueue recycle, count broadcast).

⚠️ On-device verification pending: the on-demand contract (build in
componentAtIndex, native attaches via sign, no eager append) is the
intended decoupled-list path but is NOT device-verified against the
Lynx fork's ListElement. Re-entrancy itself is safe (#214). The eager
path it replaces is known-good — see docs/list-design.md.
